+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - nothing
 
+## [0.0.25] - 2024-05-15
+### Fixed
+- blocking condition when a long running task on a node without available processing capacity. Instea
+d reduce the chance of a blocking condition by checking for capacity first.
+
 ## [0.0.24] - 2024-05-14
 ### Fixed
 - "Stream connection lost: BrokenPipeError(32, 'Broken pipe')" from RabbitMq with long running tasks